Fix crypto.js on arm
Due to our large amount of temps we also end up creating large stack frames and
thus add large constants to the stack pointer. That affects the encoding of the
immediates and MacroAssemblerARMv7 ASSERTs out for values that require
encoding.

This is unlikely to get fixed upstream and it's infact impossible to create a
testcase with JSC JIT due to the fact that it barely uses the stack frame.
I'd rather not patch the upstream file as it is a condition hard to find and
a patch easy to drop by accident. Instead this patch adds a simple workaround
that comes are low cost: Just load the immediate into a register and do the
addition.

Change-Id: Ia551a15d2f5f6243b295a9bfd19df778467189ec
Reviewed-by: Lars Knoll <lars.knoll@digia.com>