[masm] Prepare local stack frame for new upcoming exception handling
Access local temps through a newly allocated LocalsRegister instead of the
regular frame pointer register. In the new exception handling we're going to
re-enter our function in the middle and want to access the same local temps,
but we can't do that through the stack frame pointer then, because that one
will _have_ to continue to point to the local stack frame in order for
unwinding to work properly.

Also the callee saved registers are now stored right below the stack
frame pointer instead of at the bottom of the stack. This way they
can be described easily in the unwind info as always relative to the
canonical frame address.

Change-Id: I53ef6291d99396577a72ceb9246f7ca3d99e5137
Reviewed-by: Lars Knoll <lars.knoll@digia.com>