Fix crashes on ia32 with array literals
prepareVariableArguments uses copyValue to copy/initialize the
arguments for a variable arguments call. Previously it called
copyValue with only temps as source, which on ia32 is implemented
as
    (1) load temp from memory into FP reg
    (2) store FP reg into destination memory location

After Gerrit change Id5757dd7a910be13db4958da7e449172ec1b04ae the
source of copyValue can now also be an IR::Expr and hence a constant
value. On ia32 it is unfortunately not possible to load a constant
floating point value into an FP reg right away. However in this
very situation it's sufficient to store the correct VM::Value in
the destination memory address right away. This is now done using
a template specialization of copyValue.

The old code compiled because we accidentally picked the
loadDouble(const void *address) overload of MASM, with address
pointing to the IR::Expr :)

Change-Id: Ie43f9be432cfcb844cc5e706e63bd3e91e857bac
Reviewed-by: Lars Knoll <lars.knoll@digia.com>