Optimize exception throwing
Introduce a try statement in the IR that allows the back-end to jump to the try
and catch blocks directly instead of having a conditional jump for each entry.

For the case where we have no catch but only try { ... } finally { ... } we
create a synthetic catch block that sets hasException to true, to ensure that
when an exception is thrown during try we execute finally and then rethrow the
exception.

Change-Id: If4be5421f9731522beab80e843283b517d4aa41c
Reviewed-by: Lars Knoll <lars.knoll@digia.com>