Fix jump and property changes on first move when dragging a Flickable.
Fixes a regression introduced by https://codereview.qt-project.org/48690
whereby on the first mouse move the contentItem was moved the total
distance from the touch point to where the drag distance was exceeded.
For large drag thresholds this causes a noticeable jump. Reverting the
dragStartOffset portion of that change fixes the regression.

That same change was also resposible for the property update regression
https://codereview.qt-project.org/49768 attempted to fix.  That change
is also reverted as it was unnecessary and may have introduced some
regressions of its own.

Task-number: QTBUG-30032
Change-Id: I7723b459cc5a6a0731893aeb6332d00cad1bd79b
Reviewed-by: Martin Jones <martin.jones@jollamobile.com>