Fix shadowing of arguments through local variables
The declaration binding instantiation paragraph in the spec (10.5) says that
the formal parameters are entered into the environment before anything else.
Later when processing the variable declarations, CreateMutableBinding and
SetMutableBinding is only called if the HasBinding return false. That means in
the following example the "var foo" declaration does not have any effect, not
even setting foo to undefined:

function fun(foo) {
    var foo;
    return foo;
}

fun("a") === "a"

This patch implements this behaviour at compile-time by ensuring that no formal
parameter can be entered as member into the environment and thus writes to
these variables will overwrite the parameter values only.

This also advances execution on the EarleyBoyer benchmarks a little further,
which tries to re-declare a lot of parameters again as local variables.

Change-Id: Ic12d71fd39caa44169fc4b62526f03078c2513c4
Reviewed-by: Lars Knoll <lars.knoll@digia.com>