QQuickTextEdit: make the update logic work with multiple text frames
We can't assume that the text nodes are added in order since we're
iterating over all the blocks of a text frame before processing its child
text frames. The only way for all this not to collapse is to sort the text
nodes once we're done each time we're replacing/adding new ones.

Task-number: QTBUG-30349
Change-Id: Ia5d804f7f196b2348fd68fdd62a6585c189baaa4
Reviewed-by: Caroline Chao <caroline.chao@digia.com>
Reviewed-by: Jens Bache-Wiig <jens.bache-wiig@digia.com>
Reviewed-by: Alan Alpert <aalpert@blackberry.com>