Android: Support implicit import paths for assets:/ urls
The assets: scheme is not handled by QNetworkAccessManager, so it needs
to be special cased along side the qrc: scheme. We make QML recognize
it as a synchronously loaded, local file, and we make the type loader
get the correct path and check existence of the directory correctly.
This makes it possible to implicitly import the containing directory
of QML files that are bundled in the assets directory.

Task-number: QTBUG-30510

Change-Id: I048292757e509066a4af342081ec8428e2bd8af3
Reviewed-by: Paul Olav Tvete <paul.tvete@digia.com>