Doc: Un-overload the word "module"
"The Qt Quick module contains the QtQuick QML module and the Qt Quick
C++ module" is an odd way to describe things.

The concept of a "QML module" is deeply entrenched so this patch won't
touch it, but at least we can get rid of references to "C++ module"s.

Change-Id: Ib06f8e3152551901dc6016cb37507cb9608d7863
Reviewed-by: Jerome Pasion <jerome.pasion@digia.com>