Fix crash on Windows
Follow-up to commit 67b09aef34ff7997808e04194cdb92273ec76f14 about changing
the masm back-end to support returning the VM::Value in eax:edx instead
of through an invisible first pointer parameter: If the address of where
to store the return value is passed on the stack, it is the callee's
responsibility to take that one off the stack. We do that by emulating the
ret(n) instruction. However on Windows we don't want to do that because
we use the eax:edx register pair, so the correct #ifdef to use here is
RETURN_VALUE_IN_REGISTER, not ARGUMENTS_IN_REGISTERS.

Change-Id: I73ebd0688870348ed65ef91d7f2d73d0ba3ec542
Reviewed-by: Lars Knoll <lars.knoll@digia.com>