The resources property should be independent from QObject
Resources was a direct mapping of the QObject children.
This led to problems since the QObject children can also contain
other objects from other sources like attached properties.

This patch decouples resources from QObject properties
and also does not call setParent() anymore.

The special case for QQuickWindow in data_append does rely
on the fact that QObject::setParent() is called and the
inner window becomes a QObject::child of the content item.
So we keep the setParent for this special case.

The children property does not take QObject ownership either.
QObject ownership is handled by the VME.

None of the documented QML use cases should be touched by this change.

This is a cleaner solution then the ad hoc fix provided by
https://codereview.qt-project.org/#change,54677

I changed also the test. The list count now has to be exactly 4.

Change-Id: I5c119e333ee82e888aaac1da559fd63a875d08ee
Reviewed-by: Liang Qi <liang.qi@digia.com>
Reviewed-by: Shawn Rutledge <shawn.rutledge@digia.com>