Fix lookups for properties called "name" in QML
We give the function prototype the property name == "Function". The QML
scope object is created using a function template with a fallback property
handler and we inject it as activation object. Therefore a lookup for "name"
will end up as property looking up in activation object before going to the
fallback handler. The lookup for "name" will then succeed through the prototype
chain and find the FunctionPrototype's name property.

There are different ways of fixing this, including the QML side or the v8
binding implementation. But let's not give the function prototype the name
property - it doesn't appear to make much sense.

Change-Id: Icfb7bed7ef3e663ccb5f5cb189f664934264a558
Reviewed-by: Lars Knoll <lars.knoll@digia.com>