Prevent tab focus from wrapping endlessly
If there was no item that accepted focus, it would go into an
endless loop.

This also changes the default behavior of QQuickWindow. When there is not
any activeFocusItem in the whole window, it means the contentItem got
focused. The Tab/BackTab key will now focus the next item in the tab focus
chain.

Autotest is included.

Done-with: Frederik Gladhorn <frederik.gladhorn@digia.com>
Task-number: QTBUG-31344
Change-Id: I854292f89a327c493eec21969907c94aa9cfddcb
Reviewed-by: Jens Bache-Wiig <jens.bache-wiig@digia.com>
Reviewed-by: Gabriel de Dietrich <gabriel.dedietrich@digia.com>