Added QQuickWindow::setDefaultAlphaBuffer()
All QQuickWindows will render using the same OpenGL context, so
for a window to support transparency, the OpenGL context needs
to be created with support for transparency from the very start.
Therefore the application needs to call setDefaultAlphaBuffer()
before creating windows.  There are some relevant comments in
QTBUG-20768 although the bug itself is not the same use case
(it was already OK as long as the first window had a translucent
color, because of setAlphaBufferSize in QQuickWindow::setColor()).

Change-Id: I92e111c1a62c0d510821b646fd334e52254f8f57
Reviewed-by: Gunnar Sletta <gunnar.sletta@digia.com>